Bouncing Ball Animations python
Tkinter implemenation in python of a bouncing ball implementation. Can run in python IDLE like all other programs. When the user presses the q key, the program quits, if the x key is pressed, all circles will be cleared from the screen. When the user presses the + key, a ball of random size and color is created on the canvas. When the user presses the - key, one of the balls on the canvas will be removed, when the "u" key is pressed, the balls speed up by a factor of 1.2. When the "d" key is pressed, all balls slow down by the same factor.